refactor: move validation consistency helpers
